Add ClamAV 1.5 APIs and bump version to 1.5.0
Change build behavior to prioritize manually specified header
and library paths, when specified.

Change the environment variable names used when manually specifying
the header and library paths from:
- CLAMAV_SOURCE
- CLAMAV_BUILD
- OPENSSL_ROOT_DIR

To:
- CLAMAV_INCLUDE
- CLAMAV_LIBRARY
- OPENSSL_INCLUDE

Pkgconfig (unix) and Vcpkg (Windows) are now only used when the builder
does not provide explicit paths for the required files.
This is to prevent pkgconfig and vcpkg to be unintentionally used when
present on a system but a specific version of either dependency is
desired.

CLAM-2881